+++ packages/frontend/src/components/common/ResourceView/ResourceViewGrid/index.tsx
@@ -7,7 +7,12 @@ import { Anchor, SimpleGrid, Stack, Text } from '@mantine/core';
 import produce from 'immer';
 import orderBy from 'lodash/orderBy';
 import { FC, useMemo } from 'react';
-import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd';
+import {
+    DragDropContext,
+    Draggable,
+    Droppable,
+    DropResult,
+} from 'react-beautiful-dnd';
 import { Link } from 'react-router-dom';
 import { ResourceViewCommonProps } from '..';
 import { useReorder } from '../../../../hooks/pinning/usePinnedItems';
@@ -83,7 +88,7 @@ const ResourceViewGrid: FC<ResourceViewGridProps> = ({
 
     const { mutate: reorderItems } = useReorder(projectUuid, pinnedListUuid);
 
-    const handleOnDragEnd = (result: any) => {
+    const handleOnDragEnd = (result: DropResult) => {
         const { source: drag, destination: drop } = result;
         const draggedItemId = drag?.droppableId;
         if (!drop) return;
@@ -123,6 +128,7 @@ const ResourceViewGrid: FC<ResourceViewGridProps> = ({
                         <Droppable
                             droppableId={group.name}
                             isDropDisabled={!hasReorder}
+                            direction="horizontal"
                         >
                             {(dropProvided) => (
                                 <SimpleGrid